优化

+++ SMA/SmaAnalysisWebServiceTool.m
@@ -352,7 +352,6 @@ static NSString *user_acc = @"account";NSString *user_id = @"_id";NSString *user
     url = [url stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
     NSLog(@"downLoadUrl: %@",url);
     [upManager downFileWithsession:url checkSum:0 callBack:^(float progress, NSError *error) {
-//      NSLog(@"callBack==%f   error==%@  %@",progress,error,url);
         if (error) {
             if (callback) {
                 callback(progress,error);
@@ -833,8 +832,8 @@ static NSString *user_acc = @"account";NSString *user_id = @"_id";NSString *user
         NSString *str = [fmt stringFromDate:[NSDate date]];
         [quietDaArr addObject:@[str,[NSString stringWithFormat:@"%ld bpm",[responseMsg getLong:user_rate]]]];
         //        [SMADefaultinfos putKey:@"quietDaArr" andValue:quietDaArr];
-        [userInfoDic setValue:[NSString stringWithFormat:@"%ld",[responseMsg getLong:user_hi]] forKey:user_hi];
-        [userInfoDic setValue:[NSString stringWithFormat:@"%ld",[responseMsg getLong:user_we]] forKey:user_we];
+        [userInfoDic setValue:[NSString stringWithFormat:@"%.0f",[responseMsg getFloat:user_hi]] forKey:user_hi];
+        [userInfoDic setValue:[NSString stringWithFormat:@"%.1f",[responseMsg getFloat:user_we]] forKey:user_we];
         [userInfoDic setValue:[NSString stringWithFormat:@"%ld",[responseMsg getLong:user_sex]] forKey:user_sex];
         [userInfoDic setValue:[NSString stringWithFormat:@"%ld",[responseMsg getLong:user_age]] forKey:user_age];
         [userInfoDic setValue:[NSString stringWithFormat:@"%@",[responseMsg getString:user_he]] forKey:user_he];
